+++ resources/lang/en/general.php
@@ -2,12 +2,14 @@
 
 return [
     'add'                      => 'Add',
-    'blog-description'         => 'A blog by Barry van Veen about programming, PHP, Laravel Framework and related stuff.',
+    'blog-description'         => 'A blog about Laravel Framework, web development and related 
+    stuff. Written by Barry van Veen. Also some personal information and a link to my Github account.',
     'blog-rss-title'           => 'Barry\'s blog',
+    'characters'               => 'characters',
     'edit'                     => 'Edit',
     'errors-in-form'           => 'Errors in this form',
-    'homepage-subtitle'        => 'a blog about Laravel and web development',
-    'homepage-title'           => 'All articles',
+    'homepage-title'           => 'A blog about Laravel and web development',
+    'ideal-length-120-170'     => 'ideal length is between 120 and 170',
     'last-update'              => 'Last update:',
     'login-failed'             => 'Login failed',
     'luckytv-rss-description'  => 'All recent episodes of LuckyTV in a single RSS feed',